+++ .firebaserc
@@ -1,0 +1,5 @@
+{
+  "projects": {
+    "default": "pastastico-c4995"
+  }
+}
