+++ client/src/components/Advice/WorkingProgressRender.tsx
