+++ client/src/components/Hero/HeroRender.tsx
