+++ client/src/components/Hero/index.tsx
