+++ client/src/pages/About.tsx
@@ -1,0 +1,12 @@
+import WorkingProgressRender from "../components/Advice/WorkingProgressRender";
+import NavbarRender from "../components/Navbar/NavbarRender";
+function About() {
+  return (
+    <>
+      <NavbarRender />
+      <WorkingProgressRender />
+    </>
+  );
+}
+
+export default About;
