+++ client/src/pages/Home.tsx
@@ -1,0 +1,12 @@
+import NavbarRender from "../components/Navbar/NavbarRender";
+import HeroRender from "../components/Hero/HeroRender";
+function Home() {
+  return (
+    <>
+      <NavbarRender />
+      <HeroRender />
+    </>
+  );
+}
+
+export default Home;
