+++ client/src/pages/NotFound.tsx
@@ -1,0 +1,12 @@
+import NavbarRender from "../components/Navbar/NavbarRender";
+import ErrorRender from "../components/Advice/ErrorRender";
+function NotFound() {
+  return (
+    <>
+      <NavbarRender />
+      <ErrorRender />
+    </>
+  );
+}
+
+export default NotFound;
