--- pages/About.tsx
@@ -1,12 +1,0 @@
-import WorkingProgressRender from "../src/components/Advice/WorkingProgressRender";
-import NavbarRender from "../src/components/Navbar/NavbarRender";
-function About() {
-  return (
-    <>
-      <NavbarRender />
-      <WorkingProgressRender />
-    </>
-  );
-}
-
-export default About;
