--- pages/Home.tsx
@@ -1,14 +1,0 @@
-import NavbarRender from "../src/components/Navbar/NavbarRender";
-import HeroRender from "../src/components/Hero/HeroRender";
-function Home() {
-  return (
-    <div>
-      <>
-        <NavbarRender />
-        <HeroRender />
-      </>
-    </div>
-  );
-}
-
-export default Home;
