--- pages/NotFound.tsx
@@ -1,13 +1,0 @@
-import NavbarRender from "../src/components/Navbar/NavbarRender";
-import ErrorRender from "../src/components/Advice/ErrorRender";
-
-function NotFound() {
-  return (
-    <>
-      <NavbarRender />
-      <ErrorRender />
-    </>
-  );
-}
-
-export default NotFound;
