+++ server/database.js
@@ -1,0 +1,21 @@
+import mysql from "mysql2";
+import dotenv from "dotenv";
+
+dotenv.config();
+
+const connection = mysql.createConnection({
+  host: process.env.DB_HOST,
+  user: process.env.DB_USER,
+  password: process.env.DB_PASS,
+  database: process.env.DB_NAME,
+});
+
+connection.connect((err) => {
+  if (err) {
+    console.error("Error conectando a la base de datos:", err);
+    return;
+  }
+  console.log("Conectado a la base de datos Pastástico 🍝✨");
+});
+
+export default connection;
